requirements.txt
anthropic>=0.34.0
google-generativeai>=0.8.0
httpx>=0.25.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=3.17.0
pillow>=10.0.0
pytesseract>=0.3.10
pytest>=7.4.0
pytest-asyncio>=0.21.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
